+++ MongoSink/src/main/java/com/lc/plugin/sink/mongo/service/MdbService.java
@@ -371,7 +371,7 @@ public class MdbService {
 	private Number getNumber(Object value) {
 		String str=value.toString().trim();
 		if(0==str.length()) return null;
-		if(CommonUtil.isInteger(str)) return Integer.parseInt(str);
+		if(CommonUtil.isInteger(str)) return Long.parseLong(str);
 		return Double.parseDouble(str);
 	}
